Move the add_pdf_pages method to the PageAdder class
I did that factorization thinking that undo/redo would be implemented with the
command design pattern. But then I switched to the memento pattern which do
not require this change. Yet I prefer to keep it because it's still good to
factorize even if recognize this change could have been a bit simpler.