menu: Add 'new' button to open new document
I was arranging PDFs, and was going to organize multiple files into
multiple PDFs. What I was missing was a way to create a new empty
document from withit PDF Arranger; this commits adds that.

It simply works by spawing a new instance of PDF Arranger using the app
launch context from Gdk, using the .desktop file of PDF Arranger itself,
which is what I was in fact was doing to create new documents, but via
the launch mechanism in the desktop environment.