Refresh iconview on zoom change, not on thumbnail update
Iconview refresh is triggered when some parameter change (margin,
column number, spacing..). When cell is at min limit 50px, and when
cell is wider than window, none of theese will change. In theese cases
a refresh need to be triggered. This can be done by writing a page to
model. App performance is best if this is done at zoom change.