Limit max crop to 90%
Because memory usage will go very high with current 99% limit if a PDF
with a lot of pages is cropped to 1% height and width and then zoomed
in as much as possible. Mem usage will go from about 15GB to 2GB with
this commit on a 2560x1440 screen (cropped A4 pages).