Render cropped thumbnails in core
Currently the full page is rendered and crop is removed in iconview
by rounding to nearest integer. In the exported page the crops are
removed without rounding. The result is that the thumbnail and the
exported page will not look exactly the same.

To overcome that the crop has to be removed with float calculation
at rendering.

Pos: Thumbnail will be cropped as the exported page.
Pos: Memory usage is decreasing when the cropped areas are not rendered.
Cons: Page need to be re-rendered after crop.