Block drag & drop when a modal dialog is open
The main window is greyed out when a modal dialog is open so you would
not expect that dnd works. But on X11 and Windows files & pages can be
dropped into iconview. (On Wayland dnd already seems to be blocked)
If pages are dropped into iconview while a dialog is open you might
end up editing other pages than you expect. (for example)