Remove some paste as overlay/underlay limitations
Allow to paste pages with layers as layer

Allow to paste multiple pages
The copied pages will be pasted in order to the selected pages. If
number of selected pages is greater than number of copied pages,
the process will repeat, starting from first copied page.

Allow to paste pages with any size
The paste layer dialog will show the size of the first copied page and
the first "paste to" page. Percentual offsets still apply for all pages.