Compatibility switch for saving document-level information
Let user decide whether document-level information, such as outlines,
tags, encryption passwords, etc., shall be taken from the primary input
file and be preserved in the output file. This became the default
behavior with PDF Arranger 1.11 using pikepdf version 8 or newer, which
uses QPDF jobs. Prior to that, the output was always built starting with
a new, empty pdf document and thus all the document-level information
from the input file was discarded in the output. This old behavior can
now be enforced by deactivating the check button
"Keep document-level information from first page's file" in the save file
chooser dialog. If this option is deactivated, the "--empty" qpdf option
is used inplace of the primary input file.

Further, if the check button is kept activated (default), the pdf file of
the first page to be saved is used as the primary input file. Previously
the first opened file was used, which lead to unexpected/surprising behavior
if all pages of this file are removed before saving or if other files' pages
were moved to the beginning of the resulting document.