Add switch to save into empty file
Let user decide whether document-level information, such as outlines, tags,
encryption passwords, etc., shall be taken from the first opened file and
be preserved in the output file when saving or exporting to a single file.
This became the default behavior with PDF Arranger 1.11 using pikepdf
version 8 or newer, which uses QPDF jobs. Prior to that, the output was
always built starting with a new, empty pdf document and thus all document-level
information from the input file was discarded in the output.
This old behavior can now be enforced by deactivating the check button
"Retain document-level information of the first opened file" in the preferences
dialog.
This new check button is only available when working with pikepdf 8 or newer.