TST: Make dogtail test batches independent of each other
by starting each test batch with a fresh config file. Per test config
settings can be applied on top of the default config settings in
PdfArrangerTest by using applying the with_config decorator on the given
test batch.
Dogtail tests can only be run on Linux where the APPDATA environment
variable is not in use and can hence be utilized to define the config
file location for the tests.