save to original iff it's the only one in queue
Modify the behavior of "Save" to match typical meaning of that action;
specifically, overwrite the file if only one was opened.
In detail:
As before, continue to use the previously saved filename.
However, if this is the first save (and save_file is empty),
then inspect the pdfqueue.  If there is only one file open, then use
that filename.  Otherwise, as before, display the filename chooser.