Make optional the saving of window geom on exit
Introduce a config option "save-window-geometry", which defaults to
"true", and controls whether transient window properties will be written
to the config file upon exit.

By setting the option to "false" in the config file, transient window
properties will not be saved, and the config file will be modified only
if there are other (more important) preference changes.